Populate selected_functions in ClientGgp
If the user has included functions to hook in the capture, it is needed
to populate selected_functions before starting the capture.

Created ClientGgp::GetSelectedFunctions which returns the map with the
functions with the format accepted by CaptureClient. Also created two helper methods:
ClientGgp::IsSelectedFunction and ClientGgp::InformUsedSelectedCaptureFunctions

The list of selected functions acts as a "filter" so all the functions
existent in target_process_ that matches any of the strings provided
will be included in selected_functions.

The user is informed of the following through logs:
1) If any of the functions provided in the list did not get a match of a function to hook, an error will be displayed informing about them. However, the capture will start. All the functions found that did match any of the filters in the list will be included in the capture.
2) List of selected_functions included in the capture.