Move CaptureData to App
This finally removes Capture.h/Capture.cpp.
Writes to CaptureData should be thread local, thus only occur in the
CaptureListener part of OrbitApp. As we need to access it also while
capturing in real-time (e.g. for selection reports), it can also not
just be moved to process manager, once done with capturing.
There is currently, no synchronization (as there might be reads while
capturing), as there were not before.

Test: Compile, Do captures, select callstacks (while capturing), load
symbols.
Bug: http://163020637