Load symbols off the main thread
With line number symbol loading is getting slower and for
large binaries when the main thread is in suspension for long
enough the tunnel generates socket error on the next request
(it could have something to do with tcp timeout).

With recent code refactoring it now possible to do symbol loading off
the main thread.

Test: Start Orbit, load symbols see that loading is done off
      the main thread.
Bug: http://b/156460883