Add "Load Symbols" context menu action to top-down view
Bug: http://b/162707037

N.B.: Loading a module's symbols, regardless of whether it's done from the
modules table, the sampling tab, etc., causes the top-down view to be refreshed
(the same happens with the sampling reports, to take the new information into
account, e.g., to group addresses under the same function) and hence to be
collapsed (unless a search string is in place). This is not caused by this
specific change, but it makes it more evident.