Add brief explanatory comment
Added a comment in ClientGgp::OnTimer as suggested in the review. It is
necessary to check that the function related to the timer info exists in
the process. If it does not, it means something has changed in
the model and it is not right so it fails before trying to update the stats of
the function.