Add fallback sorting to SamplingReportDataView
SamplingReportDataView used to use stable_sort. stable_sort's outcome
depends on the previous state of the container. Since recent changes
trigger a reconstruction of SamplingReportDataView stable_sort
preserving behaviour doesn't work anymore since the container's previous
state is not preserved. That lead's to reorderings observed by the user.

This can be avoided by establishing a total ordering. In this commit I
use the function address as a unique and totally ordered field.

Test: Manual UI test
Bug: http://b/167541131