Fix Broken Capture Duration for Empty Captures
When opening the capture view for the first time showed "days"
instead of "0". This fixes this issue, by making EventBuffer's
callstacks_ read methods const.
Previousely there was an effective access in EventTrack GEventTracer.GetEventBuffer().GetCallstacks[tid];
which was executed always at the beginning for the all threads track
and added empty callstacks for that thread id. Additionally, the
check for the capture time span asks for EventBuffer::HasEvent, which
just checked whether the map is empty (which it was not).

This also made some minor local cleanup refactorings (nodiscard, naming).

Test: On Empty Capture: Go to Sampling View. Also clear capture.
Bug: http://b/162706447