Collect the selected kernel tracepoints events
Create a hash mechanism is order to not send the tracepoint's category and
name every time an event occurs if these do repet themselves.

The only data that will be sent from the service to the client will be
the one that differs for every single tracepoint event, like the pid,
tid, timestamp and cpu. To distinguish the category and name of the tracepoint, a hash key will be associated to these, such that these are sent only once.