Match exact version when deploying Orbit
Checking the version that's installed on the gamelet is currently
checked with a simple `grep` which leads to `1.51` matches `1.51-333-xxxxxx`.

This should not be the case, so this commit changes the matching such
that always the full version number is matched.

This is only an issue when using developement versions of Orbit, so not
a customer-facing issue.