Change EventBuffer::GetCallstackEvents to ...InTimeRangeForThreadId
Fixes a bug caused by the fact that when
`SamplingProfiler::ProcessSamples` is called for a selection, the
`CallstackData` passed to it contains each CallstackEvent twice, once
for its thread and once for the fake -1 thread. But
`SamplingProfiler::ProcessSamples` assumes the opposite.

Bug: http://b/167672728

The original version is not needed as this method is only used in one place.