Send the kernel tracepoint events from the service to the client (#1181)
* Send the kernel tracepoint events from the service to the client and
create a hash key for the the tracepoint info.
Create a hash mechanism is order to not send the tracepoint's category and name every time an event occurs if these do repet themselves. The only data that will be sent from the service to the client will be the one that differs for every single tracepoint event. To distinguish the category and name of the tracepoint, a hash key will be associated to these, such that they are sent only once.
When an event occurs, we verify whether the tracepoint name and category
has been seen before. If this is the first time, we compute the key and
send the key and the tracepoint info associated separately to the client such that the next time the client sees the key, it will be able to recognise the event.Then, the whole event is sent to the client with the key computed and without the tracepoint info. Since the client already is being able to recognise that key, it is able to correctly identify the name and category of the event.

* Compute the hash key by merging an arbitrary number of strings into one
(the tracepoint name, category and ':' to mimick how the tracepoints are
usually displayed) and produce a hash for the new generated string.