Add functionality to data classes
This adds functionality to DataManager and related classes (ProcessData,
ModuleData) that was done by OrbitProcess, OrbitModule and Pdb before.
This includes mostly lookup of absolute addresses to functions and
modules.
This also introduces the MemorySpace struct to represent a space in
memory where a module is loaded by a process.