Move OrbitApp::GetCaptureFileName to CaptureSerializer
CaptureSerializer contains the logic to save a capture, which is related
to providing the name to the file where it is saved. The method
GetCaptureFileName was in OrbitApp so it was not reusable in other
clients where it can be useful. This change moves it and makes it
available.

It has been created a new file_management namespace in
capture_serializer containing the code related to the file name. A new
method called IncludeOrbitExtensionInFile that includes the .orbit
extension to the file name if it is not included already. This method
will later be used in the ggp client to include the orbit extension in
case it is not included when the file name is provided by the user