Make CaptureDeserializer independed from OrbitGl
CaptureDesirializer now uses CaptureListener methods in order
to process the data from the capture to be loaded instead.
Prior to this, capture data was loaded and directly set at once to the App.
Also, the timers were stored directly on the time graph.

The calls to capture listener must not be done on the main thread.
Therefor, as a side-effect, deserialization does not block the UI
anymore and is somewhat "interactive", i.e. you see events poping
 up on loading.

Follow up: Make the events not appearing in blocks, which requires
 rethinking of the way we store timers.

Known Issue:
When loading a "broken" capture from the startup window
 the following happend before:
 1. An error msg. appears, that and why loading failed. In the background
    Orbit's main UI is already spawned.
 2. After clicking okay on this, the UI closes, and there is another error
    msg. saying "no such file".
 3. After clicking onkay on this, the startup windows shows up again.
Now:
 1. As before.
 2. On click of okay, you are in an empty Orbit and can try to load another
    capture from the capture view, but in order to connect to a gamelet,
    you need to restart Orbit manually.

Bug: http://b/167200870 and http://b/166094073
Test: Compile, Load Capture (also a broken one) from UI and start-up window.