Map the tracepoint events to the thread ids and time (#1197)
* Map the kernel tracepoint events to the thread ids and time
For every thread id we monitor which tracepoint events have been
triggered and at what time. For that, a red-black tree has been utilized
(std::map) that has as the key the thread id and as the value another
red-black three to map the pair time and tracepoint event. A sorted
associative container has been chosen such that the tracepoint
events are sorted: for every thread id found in an non-descending order,
the events are found in a non-descending order of time.
Every occured tracepoint event not only is it map to the corresponding
thread on which it has been activated but also, it is stored in an
universal entry dedicated to all the thread ids
(SamplingProfiler::kAllThreadsFakeTid).
At the end, the tracepoint event is registered, meaning that the upper
and lower bound for the timeline on which the events happen can be
updated.