Add orbit_service::Process type
We currently directly use the proto `orbit_grpc_protos::ProcessInfo` as
a data type in `ProcessList`. This commit adds a `Process` class which
inherits from `orbit_grpc_protos::ProcessInfo`.

This additional class allows to add additional caching data members on
top of the data coming from the proto. We are going to use to store
caching CPU usage data temporarily.