Remove top from obtaining CPU usage information
Up to now we use top for getting CPU usage information per process.
This requires a fixed output format of top and is slow since top has to
run a certain time interval for averaging CPU usage.

We can do the same internally by reading /proc/stat and
/proc/[pid]/stat. This commit adds this logic to
orbit_service::ProcessList.

The routines reading from `/proc` live in orbit_service::utils and are
unit-tested.

The CPU usage is averaged between two consecutive calls of
ProcessList::Refresh. This means the averaging time period is not
constant and depends on when Refresh is called. But I believe it's good
enough for our use case.

Test: Manual test, compared process list cpu output to top running in
shell
Bug: http://b/165823980
Fixes #1157