Properly Initialize Callbacks with Dummy Functions
Recent changes from "if (..._callback_)" to "CHECK(..._callback_)"
led to failed checks in the Fuzzers, as the callbacks were not
explicitly set.
Those callbacks were default initialized, which would lead to
an runtime exception being thrown on call.

This change initializes all callbacks with functions doing nothing,
but are save to call. Also it replaces the remaining "if"s with
"CHECK"s.

Bug: http://b/168770539
Test: Compile