Update double-conversion recipe revision in lockfiles
The lockfiles (which are used on the CI for consistent builds) still
require the build to link an old revision of the double-conversion-
package which is no longer available on the official remote (conan-center).
The build still works because we have a copy on the internal artifactory server.

This commit fixes that problem by switching to the current official
revision.