Manual instrumentation "async" operation tracking
An async operation can start in one scope and end in another,
possibly on a different thread altogether. Start and stop pairs are
matched using a user provided "id". See ASYNC_START and ASYNC_STOP
macros in Orbit.h. A new ASYNC_STRING macro allows to associate an
arbitrarily long string to an async event by specifying the event "id".

On the UI side, a new AsyncTrack takes care of displaying the async
spans in the capture view in a compact way.