Simplify and generalize TopDownNode hierarchy
Previously, this hierarchy was assuming that the top level would include only
threads, and that threads and functions would only have function children. This
was achieved with `TopDownInternalNode`.
But for a more general call tree (e.g., bottom-up) this doesn't hold anymore. As
clients of `TopDownNode` (namely, `TopDownViewItemModel`) were already checking
for the actual type anyway (`TopDownFunction` or `TopDownThread`), this
simplifies the hierarchy without changing how these classes are used.