Minor Improvements in SamplingProfiler
Prior to this, GetCallstacksFromAddress was modifying the total
callstack count, which would not be expected from a getter.
Now the caller takes care of that (same runtime costs).

Also GetSortedCallstacksFromAddress was returning a shared_ptr,
which was only stored in the SamplingReport. So unique_ptr makes
the owner ship more clearer.

Bug: http://b/166238019