Sort CallstackRepors by SampleCount
Prior to this, callstack reports were attempted to be sorted by thread usage.
However, this was basically never set (except for the all thread case).

As the thread usage is also not used anywhere else, this removes this
information and uses the sample count for sorting (as already implicitly
doe in top down view).
This also simplyfies the code.

Bug: b/169135142
Test: Take a Capture and see that the thread taps in the sampling
report are now actually sorted by number of samples.