Add QWebChannelExtractor
The JavaScript client library for Qt's WebChannel module is embedded
into the QtWebChannel shared library. For the development workflow we
need to extract that file. The simplest portable way to do that is to
compile an exectuable link it to QtWebChannel and output the resource.

On Windows where we use Qt from a conan package we could also copy the
library from the source archive, but this won't work on Linux where
we usually rely on system-installed Qt. That's why this extraction
mechanism is the simplest portable way of getting the right version of
that file.

This mechanism has a limitation: It won't work with cross-compilation
setups. But this can be easily added in the future by separating the
extractor into its own conan package and require that package as a build
requirement which means it will be compiled for the build platform and
not for the target platform. Currently we don't need the cross
compilation workflow for the UI.

Bug: http://b/169113529