Only show capture window statistics in developer mode
When pressing 'I' in the capture window a couple of debug windows appear
which are helpful for Orbit developers but not so much (at all?) for
Orbit users.

This change should prevent that users accidentally open these debug
windows by pressing I. With this change 'I' will only work when Orbit is
in developer mode. Started with the `--devmode` command line option.