Enable dev-mode via env variable
Enabling the development mode is now necessary to show the
CaptureWindow's debug statistics windows. When working on the
CaptureWindow the stats are handy all the time, so having a mechanism to
always enable them comes handy.

This change introduces an environment variable ORBIT_DEV_MODE (or
alternatively ORBIT_DEVELOPER_MODE) which when set to any non-zero
value enabled the developer mode. The command line option `--devmode`
continues to work.