Add first webpack bundling integration
This is a first step for integrating the Webpack bundler into our cmake-
and conan-based build system.

conan will provide a distribution of NodeJS as a build dependency. CMake
will then take care to install required NPM-packages during its
configuration phase.

The actual invocation of webpack takes place during the build phase and
it supports incremental builds. The incremental part is currently
mainly handled in cmake which requires that all input files are also
enlisted in the CMakeLists.txt file. This can become a source for bugs
pretty easily and maybe also an annoyance. So this can be changed to a
different approach which would make the cmake-code more complex. That's
why I went with the current (explicit) solution for now.