Fix Function Offset in Serialization (#1291)
Sampling information just contains raw absolute addresses.
Mapping those to functions (via offsets) and function names
is done via two mechanisms:
1. If available, the loaded symbols on the UI are used.
2. Otherwise, it will be checked if there are address informations
   from unwinding (the service).
On serialization, we do not store the complete information of loaded symbols (yet?),
but only save the address information, received from the service.

As we do not want to modify this information (potentially during capturing,
as loading symbols is possible while capturing), we already patch the serialized
version of the address information. However, that was limited to
writing the function name. In particual the offset was missing.

This adds this information.

Bug: https://169391773
Test: Take a capture, load symbols, save the capute, restart orbit, load capture -> symbols are correct