Make cmake_installer non-transitive build-dependency
conan has some problem with the new lock-file mechanism and the
profile-provided build dependencies. I'm not sure yet whether that is a
bug in conan or if we are doing something wrong. But I wanted to make
that change anyway for a long time.

This change tells conan that cmake_installer is only a build
requirement for the entry package, but not for dependencies. In our case
that means: "Orbit needs cmake to be built, but not its dependencies.
(These might defined their own build dependencies.)"