Fix nightly and release builds
The new version of conan logs all options to the standard output, which
was not the case before. One of the options logged in `crashdump_server`
which is filled from a secret from Keystore even though it's not so
secret.

Kokoro has a script which abort the build when something logs a
keystore-secret and that breaks the build.

This commit fixes that be adjusting conan's output with sed.