Store ProcessData by Value (Not as Unique Pointer) (#1301)
ProcessData may be stored at two places, first in DataManager and
second in CaptureData. There is no need to store them as unique
pointers here. Moreover, the absense of a concrete process in
CaptureData already led to a crash.

This change ensures, that we always store the process by value,
which among others allow clear default values for empty captures.

Test: Compile