Store Selected Functions as Values
This is the first part of a series of moving away from representing
functions by absolute addresses in order to make them independent
of the loaded process.

Prior to this, DataManager stored the selected functions as
as set of their absolute addresses.
This changes this, and stores them by value.
As protos have no hash/equals implementation, FunctionInfoSet
was added.

Test: Compile
Bug: http://b/169309553