Add --build=outdated to conan lock commands
After a series of tests I came to the conclusion that this flag is
needed when creating the full lock-file. Without it conan won't include
build requirements into the lockfile for dependencies that need to be
built locally. And when these are not included the next step (`conan
install`) will fail because it is missing a locked dependency from the
graph.