Remove FunctionInfo module_base_address
This commit removes module_base_address and load bias from FunctionInfo
(see http://b/169309553). To accomplish that a lot of files needed to be
changed. The major conceptual changes are as follows:

ModuleManager is introduced and used as a thread-safe (synchronized)
access to modules and the symbol data that modules hold. It exists only
one instance of ModuleManager at all times, which lives in OrbitApp.
CaptureData stores a pointer to this ModuleManager. This eliminates the
need to store a module map in CaptureData.

Storing and loading captures is changed. The stored capture now contains
full information about process and modules (no symbols) instead of just
process name and id.

Since FunctionInfo does not contain module base address anymore, symbol
loading is simplified and does not involve Process anymore.