Serialize Process and Modules of Process
With removing process information (load address) from functions,
this information needs to be reconstructed on loading/saving a
capture.

This fixes a crash on saving a loaded capture.

Currently, the grpc protos ProcessInfo and ModuleInfo are copy pasted
to the client protos, as the build pipeline does not properly support
integrating protos from different projects.
This will be solved, once we only rely on grpc protos.

Test: Compile & run tests
Bug: http://b/169309553