Draw white text boxes with black text for iterators overlay
The white text showing iterator timings is hard to read and can be
easily confused with other text. This is in particular true as all
text for textboxes is drawn at the end above essentially everything
else. We fix this here partly by using solid white text boxes with black
text for the iterator timings.

Since the white timer text is still drawn last, the black text for the
iterator timings is still overdrawn with timer text when text is drawn
in the same position.