Fix "node not found" problem
NodeJS is installed by conan as a build dependency but not added to the
user's global PATH variable. So the conan-supplied NodeJS executable is
only found when the executing environment is initialized by conan, i.e.
when cmake is called from conan like it is the case in build.{sh,ps1}.

When cmake is called either manually or by a third party like CLion the
NodeJS-executable won't be found because it's not in the PATH. This can
be circumvented by caching the contents of the PATH variable during the
configuration phase and restoring it on every subsequent call of cmake.

Limitation: This requires the initial build to be initiated by conan,
usually by running `build.{sh,ps1}` or the bootstrap script.