Add CodeView web target
This adds a first version of the code view
to our webpack pipeline. It depends on PrismJS for
syntax highlighting. The HTML entry point is created
from a handlebars template.

The page relies on QWebChannel for communication and
is intended to be loaded in a dialog window inside Orbit.
Code for the C++ part will follow in another commit.

It also contains a custom heatmap plugin which allows
to show per-line-call statistics next to every source line.