Only request data update on live view when needed
The live functions data view is handled differently than other data
views as creation of it is a bit more complex than for the view we can
handle generically. Therefore, this data view requires an explicit
extra call to OnDataChanged. So far, this call was done independent of
the type of data views to be updated. This leads to a bug with frame
tracks where some data in the live functions data view class is reset
that should not be reset. Note that this mimics the update behavior of
the generic other data view panels.

Bug: b/170392477.
Test: Manual test of taking a capture, adding and removing frame
track(s).