Add test isolation to gtest
By using Android's rewritten main function for gtest
we can support test isolation, meaning each test in a
test binary will run in its own process.

This feature is not enabled by default but can be enabled
with the `TEST_ISOLATION` flag for the cmake `register_test`
function.

Furthermore this change adds test timeout support. The default
timeout will be 30 secs and can be changed by the `TIMEOUT` option,
that needs to be passed into the `register_test` cmake function.