Restrict Disassembly Action to one function
Previously it was possible to select multiple functions and click "Go to
Disassembly" and Orbit would open as many Disassembly windows as needed.

This workflow is not (yet) supported by the web-based Disassembly View
and we also want to avoid that the user has to juggle with multiple
dialog windows. This will be solved differently in the future, probably
with something like tabs inside a single window.