Split web_engine::Dialog's responsibilities
There are now two classes View and Dialog. Dialog handles the
QWebEnginePage while View provides the web channel and the
WebSocketServer.

This way the Dialog can be reinstantiated independently from the View.
That avoids problems with old state spilling into the next use of a
Dialog instance. The webpage will always be reloaded, scroll positions,
focus elements are all reset.

The term View is used here in the way it's defined in Orbit, not how
it's defined in Qt. A View is the data source for the webpage in a
dialog.