Add Vulkan Layer Skeleton Including Bootstrap Code
This add the skeleton code for a vulkan layer that should track
the GPU time of command buffers.

It contains all the bootstrap code needed, as well as some basic
tracking of command buffers.
What's missing though, is the actual insertion of timestamps + all
the management that comes with that.

Test: Compile
Bug: http://168200502