Fix incorrect computation of the maximum box ratio for frame tracks
Maximum box ratio determines the scaling factor for the boxes on the
track and is based on the largest box height and the average box height.
So far, this ratio was computed online whenever a timer was added. This
is incorrect as the average value (obviously) still changes with future
timers to be added.

We fix this be computing the maxium scale factor (box ratio) only when
needed for drawing when all timers have been added.

This approach should also be correct when live updating a frame track
during a capture, which is not working yet, but planned for the future.