Remove package upload from CI build script
We used to automatically upload built packages to the internal
artifactory server when they have been built by the CI during the
run.

But with build_and_upload_dependencies.sh we have a better solution
for a while now, which we have been using exclusively for uploading
new packages.

Furthermore this upload step once corrupted the cache in the past due
to a partially uploaded package.

This all concludes that the best option is to remove the automatic
package uploading and require manual uploading with
build_and_upload_dependencies.sh.