Introspection revival
We can now use the manual instrumentation API from OrbitService and
instrument our own code (and graph variables) alongside the target
process data in the capture view. Introspection tracks are displayed
right after the scheduler track and have a darker background. The
feature is automatically enabled with the "--devmode" flag.

- Use orbit_api's EncodedEvent class for introspection data
- Properly set pid on uretprobes
- Add new IntrospectionScope capture event
- Draw track background based on track pid
- Instrument TracerThread using introspection