Track ordering basics (#1365)
Basics for thread pinning and sorting.

Pinned tracks are always displayed on top of the capture window,
similar to frozen panes in a spreadsheet. Scrolling will not affect the
position of those tracks, and other tracks are rendered underneath
them.

Track tabs can be clicked and dragged up and down to adjust the
sorting in the capture window.

Core changes

Changes are distributed across multiple commits. Basically:

Removed unneeded definition of methods in Pickable (now defined
by the tracks)
Added z_offset parameter to all drawing / update functions to
allow stacking thacks "on top" of others
Auto-sorting of tracks is now only happening during a running capture
Open issues

Sorting / Pinning is not saved with the capture, and lost when the
capture is restarted
Not all elements are ordered correctly w.r.t. depth (e.g. rounded
corners pose an issue)
Due to the open issues, this change is behind a feature flag.
Use -track_ordering_feature to enable.