Add automated test for frame tracks
This adds an automated test for creating of frame tracks. Since this
requires hooking a function and taking a capture first, the whole
process of hooking and capturing is refactored to be reusable between
the test for instrumenting a function and the new frame tracks test.

Tested: Ran all automated tests locally.