Add automated test for iterators
This adds an automated test for iterators. The test hooks a single
function, then adds an iterator for the function. Verification is done
by comparing the number of buttons before and after addition of the
iterator, as we know that a single iterator adds three buttons
(previous, next, and delete).

Tested: Ran locally.