Skip events coming before all file descriptors have been enabled
Enabling the perf_event_open file descriptors can take a while when we have a
lot of them (and also once we go back to allowing to start a new capture while
the file descriptors from the previous ones are still being closed).
Some events can start coming to the ring buffers from the first file
descriptors to have been enabled, while we are still enabling other file
descriptors and hence before we start reading from the ring buffers.
`PerfEventProcessor` would have to buffer for all this arbitrarily long interval
in order to guarantee events to be processed in order.
Therefore, we discard events that came before the last file descriptor had been
enabled.
We prefer this to simply enabling all file descriptors in a different thread, as
we want to perform additional work after enabling the file descriptors but
before starting to read from the ring buffers.
This also prevents the beginning of the capture to be missing parts of
information as some events are already being recorded and others aren't.
Also refer to http://b/170711685#comment4.

Bug: http://b/170711685

Test: Capture Trata hooking all the `vk*` functions in `libvulkan.so.1`,
see from the logs that no event is out of order.