Close perf_event_open file descriptors in a separate detached thread
Closing those file descriptors can take a long time, but all data has already
been correctly processed and sent.
Note that if a new capture is started while those file descriptors are still
being closed, it can take a while for that capture to be effectively started and
for data to come in. But this is the behavior we had when the file descriptors
were closed by the forked `top` process (see explanation in
http://b/170711685#comment17). So let's quickly fix this regression.

Bug: http://b/171350567

Test: Both locally and on Trata, try taking multiple captures in a row when
instrumenting a relatively large number of functions. Verify that the behavior
is like the one prior to the removal of `top`.