Preparation for frame track serialization
Data for frame tracks is now kept in App and that is the source of
truth for added frame tracks. This is in preparation of supporting
serialization of frame tracks (and possibly live updates during a
capture, or at least maintaining added frame tracks in-between
captures).

This change fixes a bug on master where removing a frame track was not
possible anymore.