Use ClientContext::TryCancel to actually abort the capture when exiting
So far, "aborting" the capture on exit only meant stopping it, which could still
take a long time. Now this is immediate.

This even seems to allow to close Orbit in the situation of http://b/171923841.

Test: In addition to several local and normal captures on Trata, also capture
Trata as per http://b/171923841#comment17 until the issue arises, and try to
(now successfully) close Orbit.