Add proper QObject hierarchy to OrbitSshQt
All member variables of ServiceDeployManager that inherit from QObject
need to have ServiceDeployManager as parent.

This is needed to be able to change the thread association of all
ServiceDeployManager related objects in one go.

When calling "QObject::moveToThread" on a ServiceDeployManager instance
all QObject-children will also change its thread association and related
events will be processed on that other threads event loop automatically.