Use sched:sched_switch tracepoint for context switches
This replaces the use of `perf_event_open`'s built-in
`PERF_RECORD_SWITCH_CPU_WIDE`, as the `sched:sched_switch` is already used for
thread state.

This solves the issue of scheduling slices and thread states not being
perfectly aligned as corresponding `PERF_RECORD_SWITCH_CPU_WIDE` records and
`sched:sched_switch` events have slightly different timestamps (by a few
microseconds). And it avoids the overhead of using both.

Furthermore, while with `PERF_RECORD_SWITCH_CPU_WIDE` events we had one event
for the switch out and one for the switch in with slightly different timestamps
(see `perf_event_task_sched_in` and `perf_event_task_sched_out` in
https://github.com/torvalds/linux/blob/master/kernel/sched/core.c#L3616), which
always caused a gap between consecutive scheduling slices (the time spent in the
scheduler), with `sched:sched_switch` tracepoints we have a single event for
each context switch. This means that a gap between consecutive scheduling slices
is present only if the CPU is actually idle in that time, which is clearer.

The additional complexity comes from making sure that the process id be
available for every scheduling slice, including in particular when a context
switch is caused by a thread exit.
For example, now context switches always go through `PerfEventProcessor`, as
this is already required by thread state tracking and as they need to be
processed in order with `PERF_RECORD_FORK` to keep track of the process ids.
This has the advantage of moving such processing out of the thread that reads
from the ring buffers. But as expected, the additional processing means that the
overall performance *when not collecting thread states* gets worse.
The difference in CPU utilization of `OrbitService` is not dramatic but in
relative terms it can be high when none or only "light" dynamic instrumentation
is done.
But in the context of collecting thread states (which obviously require more
CPU usage as more data is collected) this is of course a performance
improvement (as they already use `sched:sched_switch` tracepoints, see above).

Bug: http://b/168790833

Quite some testing on Trata and Infiltrator to verify that everything looks
good in parallel with performance measurements.